Fix cascade sqlite not working

+++ app/src/lib/cascades.js
@@ -3,7 +3,7 @@ import * as FileSystem from 'expo-file-system';
 import * as SQLite from 'expo-sqlite';
 import { conn, query } from '../database';
 
-const DIR_NAME = 'Cascades';
+const DIR_NAME = 'SQLite';
 
 const createSqliteDir = async () => {
   /**
@@ -37,8 +37,10 @@ const loadDataSource = async (source) => {
 const dropFiles = async () => {
   const Sqlfiles = await FileSystem.readDirectoryAsync(FileSystem.documentDirectory + DIR_NAME);
   Sqlfiles.forEach(async (file) => {
-    const fileUri = FileSystem.documentDirectory + `${DIR_NAME}/${file}`;
-    await FileSystem.deleteAsync(fileUri);
+    if (file.includes('sqlite')) {
+      const fileUri = FileSystem.documentDirectory + `${DIR_NAME}/${file}`;
+      await FileSystem.deleteAsync(fileUri);
+    }
   });
   return Sqlfiles;
 };
